[CAP-001] Refactor, InitialVC, VM 이름변경및 구조 리팩토링
