+++ Modules/Presentation/MainAppFeatures/Source/OnBoarding/InitialScreen/RootVM.swift
@@ -10,21 +10,41 @@ import Entity
 import RxSwift
 import RxCocoa
 import UseCase
+import BaseFeature
 import Util
 
-public class InitialVM: InitialViewModelable {
+public protocol RootViewModelable: BaseVMable {
     
+    var coordinator: RootCoordinator? { get set }
+    
+    // Input
+    var viewDidLoad: PublishRelay<Void> { get }
+    
+    // Output
+    
+}
+
+public class RootVM: RootViewModelable {
+    
+    public weak var coordinator: RootCoordinator?
+    
+    // Input
     public var viewDidLoad: RxRelay.PublishRelay<Void> = .init()
     
-    public var tokenFlowNextable: RxCocoa.Driver<Void>
+    // Output
     public var alert: RxCocoa.Driver<Entity.CapAlertVO>?
     
     // Init
     let authUseCase: AuthUseCase
     
-    public init(authUseCase: AuthUseCase) {
+    public init(
+        coordinator: RootCoordinator,
+        authUseCase: AuthUseCase
+    ) {
+        self.coordinator = coordinator
         self.authUseCase = authUseCase
         
+        // MARK: 토큰 플로우
         let checkExistingMemberResult = viewDidLoad
             .map { [authUseCase] _ in
                 authUseCase.checkIsExistingMemeber()
@@ -47,12 +67,16 @@ public class InitialVM: InitialViewModelable {
         let tokenGenerateSuccess = tokenGenerationResult.compactMap { $0.value }
         let tokenGenerateFailure = tokenGenerationResult.compactMap { $0.error }
         
-        tokenFlowNextable = tokenGenerateSuccess
+        let tokenFlowFinishSuccessFully = tokenGenerateSuccess
             .map { _ in
                 printIfDebug("✅ 토큰 생성에 성공했습니다.")
             }
             .asDriver(onErrorJustReturn: ())
         
+        // MARK: 카테고리 확인 플로우
+        
+        
+        // 토큰 생성 실패
         alert = tokenGenerateFailure
             .map { authError in
                 return CapAlertVO(
